JavaScript Practice
Furthered understanding of how we can use the HTML DOM to alter HTML elements in JS. We even change CSS styles right in JavaScript. In addition, we learned about forms and different methods that will react to certain inputs such as the required attirbute in an input tag. Next we will began making animations in JavaScript with HTML DOM.